+++ src/interpreter/plugin/NumericAggregationPlugin.ts
@@ -114,14 +114,8 @@ export class NumericAggregationPlugin extends FunctionPlugin {
     'VAR.S': {
       method: 'vars',
     },
-    'VAR': {
-      method: 'vars', //alias of VAR.S
-    },
     'VAR.P': {
       method: 'varp',
-    },
-    'VARP': {
-      method: 'varp', //alias of VAR.P
     },
     'VARA': {
       method: 'vara',
@@ -132,14 +126,8 @@ export class NumericAggregationPlugin extends FunctionPlugin {
     'STDEV.S': {
       method: 'stdevs',
     },
-    'STDEV': {
-      method: 'stdevs', //alias of STDEV.S
-    },
     'STDEV.P': {
       method: 'stdevp',
-    },
-    'STDEVP': {
-      method: 'stdevp', //alias of STDEVP
     },
     'STDEVA': {
       method: 'stdeva',
@@ -150,6 +138,13 @@ export class NumericAggregationPlugin extends FunctionPlugin {
     'SUBTOTAL': {
       method: 'subtotal',
     }
+  }
+
+  public static aliases = {
+    'VAR': 'VAR.S',
+    'VARP': 'VAR.P',
+    'STDEV': 'STDEV.S',
+    'STDEVP': 'STDEV.P',
   }
 
   /**
